+++ src/Pages/Page/Pages.jsx
@@ -1,6 +1,6 @@
 import React from 'react'
 
-import Page from '../../Components/Page/Paje'
+import Page from '../../Components/Page/Page'
 
 export default function Pajes() {
   return (
